Configure Spec.
The approach taken here is one I use frequently. I like factorybot + ffaker as a tool that allows for randomized testing as-needed (and static values when it's not). I also added support for defining custom spec functions in a clean, manageable way. I've spent too much time in codebases without clear organizational patterns for macros, and it helps everybody when spec is very clear.